v2.0.0 #1 Fixing a couple warnings/errors for OSX beta build

+++ Xenakios/MixerActions.cpp
@@ -147,7 +147,7 @@ void DoTraxPanLaw(COMMAND_T* t)
 
 void DoTraxRecArmed(COMMAND_T* t)
 {
-	int iRecArm = t->user;
+	int iRecArm = (int)t->user;
 	for (int i = 0; i < GetNumTracks(); i++)
 	{
 		MediaTrack* CurTrack = CSurf_TrackFromID(i+1, false);
@@ -159,7 +159,7 @@ void DoTraxRecArmed(COMMAND_T* t)
 
 void DoBypassFXofSelTrax(COMMAND_T* t)
 {
-	int iBypass = t->user;
+	int iBypass = (int)t->user;
 	for (int i = 0; i < GetNumTracks(); i++)
 	{
 		MediaTrack* CurTrack = CSurf_TrackFromID(i+1, false);
